Introduce snapshots to improve compaction
A snapshot is just a db_header that contains a unique id, a pointer to
the header and a pointer to the previous header. The snapshots thus form
a linked list. When compaction runs the list is traversed to find the
oldest snapshot and compaction is performed in chunks, a snapshot at a
time walking back up the list. This allows docs that have have not been
edited to be copied only once. If compaction crashes or is restarted to
top off, it begins at the next snapshot after the update_seq.

BugzID:12639